requirements.txt
PyYAML>=6.0
watchdog>=3.0.0
pathlib2>=2.3.0; python_version < "3.4" 